### 2021-12-21 #4019 @whedon
[PRE REVIEW]: nsink: An R package for flow path nitrogen removal estimation
**Submitting author:** @jhollist (<a href="http://orcid.org/0000-0002-9254-9740">Jeffrey W Hollister</a>)
**Repository:** <a href="https://github.com/USEPA/nsink" target ="_blank">https://github.com/USEPA/nsink</a>
**Version:** v1.1.0
