### 2021-12-29 #4019 @jhollist
@arfon those two invalid DOIs are correct as far as I can tell.  The resolve with https://doi.org/ so I will leave them as is.
