### 2021-12-31 #4019 @Bisaloo
> @whedon invite @Bisaloo as editor
>
> wave @Bisaloo – would you be willing to edit this submission for JOSS?

Hi, I'm currently on leave with a limited internet access until January 20. I think it's better if you are able to find another editor for this submission.
