### 2022-01-07 #4039 @whedon
[REVIEW]: nsink: An R package for flow path nitrogen removal estimation
**Submitting author:** @jhollist (<a href="http://orcid.org/0000-0002-9254-9740">Jeffrey W Hollister</a>)
**Repository:** https://github.com/USEPA/nsink
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
