### 2022-01-11 #4039 @jhollist
@ldecicco-USGS sorry for the confusion, bit of a long story that I will spare you as to why I use the two...

It shouldn't matter.  I have my local repo set to push to both locations so, in theory, they are identical.  I use the EPA one in the paper and in DESCRIPTION so install from that.  I do see now that I need to update my README installation instructions.  Will do that shortly.
