### 2022-01-11 #4039 @ldecicco-USGS
The package works as advertised, and the paper is clear to read. I don't see specific instructions on Community guidelines (I think this is generally a little different than the code of conduct that is included).

General thoughts, not required for approval, but might help some future developments:
The tests look to me to be written to check generally that the data coming back from the functions are set up in the intended structure and without errors. Since the package is heavily based on the Kellogg paper, it would be neat to see the tests recreate and match the results from the paper.
