### 2022-01-20 #4039 @crvernon
:mega: Mid-week rally!

👋  @jhollist - It looks like @ldecicco-USGS has requested a community guidelines section be added and has made several other helpful comments.

👋 @jmp75 - Were you able to initialize your review?  Let me know if you have any questions.

Keep up the good work!
