### 2022-01-21 #4039 @jhollist
@ldecicco-USGS and @crvernon thanks for the comments!  There is a CONTRIBUTING.md but it is a little hidden inside of the .github folder.  This is the defualt location for `usethis::use_tidy_contributing` which is what I used as an initial template for my contributing.  This is one of the standard locations suggested by GitHub: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file.  If JOSS has a different suggestion I am happy to move it.

Also, I will create an issue for a future enhancement that uses the Kellogg et al data for tests.  This is a MOST excellent idea.  Just will take some time to implement.

Thanks again!
