### 2022-01-21 USEPA/nsink#2 @jhollist
Use Kellogg et al examples for tests
Based on https://github.com/openjournals/joss-reviews/issues/4039#issuecomment-1010141897

Great idea to expand tests for nsink that match the examples in https://doi.org/10.1016/j.ecoleng.2010.02.006
