### 2022-01-21 #4039 @ldecicco-USGS
Ack, sorry for not realizing that! Sure enough, it does make sense to put it in there.
