### 2022-01-21 #4039 @jhollist
I will openly admit that I googled, "why is my contributing.md in the .github folder" prior to responding!
