### 2022-02-28 #4039 @crvernon
@jhollist Please see the following concerning the paper submission:

- [ ] Please fix invalid DOI `10.1641/0006-35682002052[0129BSIPGO]2.0.CO2 URL is INVALID` in the references
- [ ] Line 11.  Missing parenthesis

❗ Please remove your old license (https://github.com/USEPA/nsink/blob/main/LICENSE) as is not valid for JOSS submissions.

Additional:
- [ ] I think you could simply reference your CONTRIBUTING.md file in your README (see the following that was an approach that was used in a recent JOSS submission:  https://gitlab.com/picos-api/picos/-/blob/master/README.rst
- [ ] Your package could benefit from a docs site.  Since your docstrings are in good shape, something like Pkgdown may be an easy implementation.  Your call on this though.

Let me know once you fix the above and we will move along.  Thanks!
