### 2022-03-01 #4039 @jhollist
@crvernon Will get this finished up this week.

One quick question.  The DOI for that reference is: 10.1641/0006-3568(2002)052[0129:BSIPGO]2.0.CO;2  and it resolves correctly at https://doi.org/10.1641/0006-3568(2002)052[0129:BSIPGO]2.0.CO;2.  It appears that some of the special characters are getting lost.  In the article proof, the square bracket are lost, but the URL resolves.  As the DOI is correct in the .bib file, I am not sure what to do to get the JOSS DOI check to pass.

I will work on a Pkgdown site for it.  It will have to be off my personal page though and not the USEPA repository as we do not have the ability to use github pages there.
