+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   intro: intro.html
-last_built: 2022-03-08T19:06Z
+last_built: 2022-03-08T19:20Z
 urls:
   reference: https://jhollist.github.io/nsink/reference
   article: https://jhollist.github.io/nsink/articles
