### 2022-03-08 #4039 @jhollist
@crvernon, first thanks for keeping this moving along.

I've made the edits, added mention of the CONTRIBUTING to the README, and have created a pkgdown site which is available via https://jhollist.github.io/nsink.  I do not have admin rights on the https://github.com/usepa/nsink repository so haven't added this link to the About section.  I can add a link to the README if you think that is good.  FYI, I should have admin rights at some point in the next few months to that repo and can make changes then.

What are the next steps?
