### 2022-03-08 #4039 @crvernon
@jhollist  we are almost there!  Next is just setting up the archive for your new release.

- We want to make sure the archival has the correct metadata that JOSS requires.  This includes a title that matches the paper title and a correct author list.

So here is what we have left to do:
- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the `main` and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Please list the DOI of the archived version here
