### 2022-03-09 #4039 @crvernon
NOTE:  the invalid DOI resolves correctly per the author's comment:

> The DOI for that reference is: 10.1641/0006-3568(2002)052[0129:BSIPGO]2.0.CO;2 and it resolves correctly at [https://doi.org/10.1641/0006-3568(2002)052[0129:BSIPGO]2.0.CO;2](https://doi.org/10.1641/0006-3568(2002)052%5B0129:BSIPGO%5D2.0.CO;2). It appears that some of the special characters are getting lost. In the article proof, the square bracket are lost, but the URL resolves. As the DOI is correct in the .bib file, I am not sure what to do to get the JOSS DOI check to pass.
