### 2022-03-09 #4039 @jhollist
Thank you, @crvernon, @ldecicco-USGS, and @jmp75!  JOSS is one of my favorite journals to submit to and review for!
